+++ display/download.go
@@ -142,7 +142,7 @@ func downloadURL(u string, resp *gemini.Response) {
 		progressbar.OptionShowCount(),
 		progressbar.OptionSpinnerType(14),
 	)
-	bar.RenderBlank()
+	bar.RenderBlank() //nolint:errcheck
 
 	savePath, err := downloadNameFromURL(u, "")
 	if err != nil {
@@ -200,9 +200,9 @@ func downloadPage(p *structs.Page) (string, error) {
 	var err error
 
 	if p.Mediatype == structs.TextGemini {
-		savePath, err = downloadNameFromURL(p.Url, ".gmi")
+		savePath, err = downloadNameFromURL(p.URL, ".gmi")
 	} else {
-		savePath, err = downloadNameFromURL(p.Url, ".txt")
+		savePath, err = downloadNameFromURL(p.URL, ".txt")
 	}
 	if err != nil {
 		return "", err
@@ -261,13 +261,12 @@ func getSafeDownloadName(name string, lastDot bool, n int) (string, error) {
 		if lastDot {
 			ext := filepath.Ext(name)
 			return strings.TrimSuffix(name, ext) + "(" + strconv.Itoa(n) + ")" + ext
-		} else {
-			idx := strings.Index(name, ".")
-			if idx == -1 {
-				return name + "(" + strconv.Itoa(n) + ")"
-			}
-			return name[:idx] + "(" + strconv.Itoa(n) + ")" + name[idx:]
 		}
+		idx := strings.Index(name, ".")
+		if idx == -1 {
+			return name + "(" + strconv.Itoa(n) + ")"
+		}
+		return name[:idx] + "(" + strconv.Itoa(n) + ")" + name[idx:]
 	}
 
 	d, err := os.Open(config.DownloadsDir)