+++ display/modals.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/dustin/go-humanize"
+	humanize "github.com/dustin/go-humanize"
 	"github.com/gdamore/tcell"
 	"github.com/makeworld-the-better-one/amfora/config"
 	"github.com/spf13/viper"
@@ -154,10 +154,13 @@ func modalInit() {
 
 // Error displays an error on the screen in a modal.
 func Error(title, text string) {
-	// Capitalize and add period if necessary - because most errors don't do that
-	text = strings.ToUpper(string([]rune(text)[0])) + text[1:]
-	if !strings.HasSuffix(text, ".") && !strings.HasSuffix(text, "!") && !strings.HasSuffix(text, "?") {
-		text += "."
+	if text == "" {
+		text = "No additional information."
+	} else {
+		text = strings.ToUpper(string([]rune(text)[0])) + text[1:]
+		if !strings.HasSuffix(text, ".") && !strings.HasSuffix(text, "!") && !strings.HasSuffix(text, "?") {
+			text += "."
+		}
 	}
 	// Add spaces to title for aesthetic reasons
 	title = " " + strings.TrimSpace(title) + " "
@@ -265,6 +268,7 @@ func Tofu(host string, expiry time.Time) bool {
 	}
 	yesNoModal.GetFrame().SetTitle(" TOFU ")
 	yesNoModal.SetText(
+		//nolint:lll
 		fmt.Sprintf("%s's certificate has changed, possibly indicating an security issue. The certificate would have expired %s. Are you sure you want to continue? ",
 			host,
 			humanize.Time(expiry),