+++ display/tab.go
@@ -62,13 +62,13 @@ func makeNewTab() *tab {
 		if key == tcell.KeyEsc {
 			// Stop highlighting
 			bottomBar.SetLabel("")
-			bottomBar.SetText(tabs[tab].page.Url)
+			bottomBar.SetText(tabs[tab].page.URL)
 			tabs[tab].clearSelected()
 			tabs[tab].saveBottomBar()
 			return
 		}
 
-		if len(tabs[tab].page.Links) <= 0 {
+		if len(tabs[tab].page.Links) == 0 {
 			// No links on page
 			return
 		}
@@ -82,10 +82,10 @@ func makeNewTab() *tab {
 			linkN, _ := strconv.Atoi(currentSelection[0])
 			tabs[tab].page.Selected = tabs[tab].page.Links[linkN]
 			tabs[tab].page.SelectedID = currentSelection[0]
-			followLink(tabs[tab], tabs[tab].page.Url, tabs[tab].page.Links[linkN])
+			followLink(tabs[tab], tabs[tab].page.URL, tabs[tab].page.Links[linkN])
 			return
 		}
-		if len(currentSelection) <= 0 && (key == tcell.KeyEnter || key == tcell.KeyTab) {
+		if len(currentSelection) == 0 && (key == tcell.KeyEnter || key == tcell.KeyTab) {
 			// They've started link highlighting
 			tabs[tab].page.Mode = structs.ModeLinkSelect
 
@@ -102,7 +102,7 @@ func makeNewTab() *tab {
 			// There's still a selection, but a different key was pressed, not Enter
 
 			index, _ := strconv.Atoi(currentSelection[0])
-			if key == tcell.KeyTab {
+			if key == tcell.KeyTab { //nolint:gocritic
 				index = (index + 1) % numSelections
 			} else if key == tcell.KeyBacktab {
 				index = (index - 1 + numSelections) % numSelections
@@ -153,10 +153,10 @@ func (t *tab) hasContent() bool {
 	if t.page == nil || t.view == nil {
 		return false
 	}
-	if t.page.Url == "" {
+	if t.page.URL == "" {
 		return false
 	}
-	if strings.HasPrefix(t.page.Url, "about:") {
+	if strings.HasPrefix(t.page.URL, "about:") {
 		return false
 	}
 	if t.page.Content == "" {